% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGO-package.R
\docType{package}
\name{autoGO-package}
\alias{autoGO}
\alias{autoGO-package}
\title{autoGO: Auto-GO: Reproducible, Robust and High Quality Ontology Enrichment Visualizations}
\description{
Auto-GO is a framework that enables automated, high quality Gene Ontology enrichment analyses visualizations. It also features a handy wrapper for Differential Expression analysis around the DESeq2 package (Love et. al 2014) and a handy wrapper around the single-sample implementation of the GSEA algorithm (ssGSEA), useful for low-numerosity transcriptional profiles. The whole framework is structured in different, independent functions, in order to let the user decide which steps of the analysis perform and which plot to produce.
}
\author{
\strong{Maintainer}: Lorenzo D'Ambrosio \email{lorenzo.dambrosio@ifo.it} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
