% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGO.R
\name{autoGO}
\alias{autoGO}
\title{autoGO}
\usage{
autoGO(
  list_of_genes,
  dbs = c("GO_Molecular_Function_2021", "GO_Cellular_Component_2021",
    "GO_Biological_Process_2021", "KEGG_2021_Human"),
  my_comparison,
  ensembl = F,
  excel = F,
  where_results = "./",
  outfolder = "results/",
  my_autoGO_dir = "~/"
)
}
\arguments{
\item{list_of_genes}{it can be a list of dataframe (i.e. the output of read_gene_list()), a vector or a dataframe containing a list of genes or a path to the .txt file.}

\item{dbs}{Databases over which the enrichment will be performed, based on the enrichR libraries. Default are GO_Molecular_Function_2021, GO_Cellular_Component_2021, GO_Biological_Process_2021, KEGG_2021_Human.}

\item{my_comparison}{Name of the comparison the user would like to inspect.}

\item{ensembl}{(Default = FALSE). Set to TRUE if the provided gene list contains Ensembl IDs. A conversion to HGNC will be performed.}

\item{excel}{(Default = FALSE). Set to TRUE if you want to save output tables in .xlsx format.}

\item{where_results}{Specify the folder in which you want to save the outputs. Default is "./". Note: if you are working with RNotebook the default working directory, if not specified, is the folder in which the .Rmd file is saved.}

\item{outfolder}{The name to assign to the folder in which outputs are saved. Default is: "results/". NOTE: please add "/" at the end.}

\item{my_autoGO_dir}{where have you cloned the auto-go repository, default is your home directory "~/"}
}
\description{
Perform enrichment analysis on all the desired gene lists. This function take advantage of the 'enrichR' package.
}
