% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comparisons}
\alias{comparisons}
\title{Comparisons.}
\format{
A data frame with 6 rows and 2 variable:
}
\usage{
comparisons
}
\description{
TODO: write a longer description
}
\keyword{datasets}
