% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conversion_ensembl}
\alias{conversion_ensembl}
\title{Mapping between gene names and Ensembl IDs}
\format{
A data frame with 52123 rows and 2 variables:
\describe{
\item{external_gene_name}{gene name}
\item{ensembl_gene_id}{ensembl gene identifiers}
}
}
\usage{
conversion_ensembl
}
\description{
Long description
}
\keyword{datasets}
