% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conversion_ensembl_hgnc}
\alias{conversion_ensembl_hgnc}
\title{conversion_ensembl_hgnc TODO}
\format{
A data frame with 52123 rows and 2 variables:
\describe{
\item{external_gene_name}{gene name}
\item{ensembl_gene_id}{ensembl gene identifiers}
}
}
\usage{
conversion_ensembl_hgnc
}
\description{
Long description
}
\keyword{datasets}
