% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrichR_database}
\alias{enrichR_database}
\title{enrichR_database TODO}
\format{
A data frame with 179 rows and 1 variables:
\describe{
\item{DATASETS}{dataset name}
}
}
\usage{
enrichR_database
}
\description{
Long description
}
\keyword{datasets}
