% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_length}
\alias{gene_length}
\title{Mapping between gene names and lengths}
\format{
A data frame with 51180 rows and 2 variables:
\describe{
\item{external_gene_name}{gene name}
\item{width}{width, in base pairs}
}
}
\usage{
gene_length
}
\description{
Long description
}
\keyword{datasets}
