% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolliGO.R
\name{lolliGO}
\alias{lolliGO}
\title{lolliGO}
\usage{
lolliGO(
  enrich_table,
  my_comparison = NULL,
  where_results = "./",
  outfolder = "results/"
)
}
\arguments{
\item{enrich_table}{Dataframe containing the enrichment results or a path to your .tsv file containing the enrichment results. Columns 'Term' and 'Adjusted.P.Value' are required. It has to contain also the column Overlap which indicates the number of genes in our list enriched over the total genes of that term (format n/N).}

\item{my_comparison}{Name of the comparison the user would like to inspect.}

\item{where_results}{Specify the folder in which you want to save outputs. (Default = "./"). Note: if you are working with R Notebooks the default working directory (if not specified) is the folder in which the .Rmd is saved.}

\item{outfolder}{The name to assign to the folder for output saving. (Default = "results/"). NOTE: please add "/" at the end.}
}
\description{
The function lolliGO.R implement the lollipop plot of the first 20 enrichment terms.

For each enrichment result table a lollipop plot is produced. Results are stored in the "enrichment_plots" subfolder for each comparison.
}
