% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene_list.R
\name{read_gene_list}
\alias{read_gene_list}
\title{Read genes list}
\usage{
read_gene_list(
  gene_lists_path = "./results",
  log2FC_threshold = 0,
  padj_threshold = 0.05,
  which_list = c("up_genes", "down_genes", "up_down_genes", "everything"),
  from_DE_analysis = TRUE,
  files_format = ".txt"
)
}
\arguments{
\item{log2FC_threshold}{Threshold value for log2(Fold Change) for considering genes as differentially expressed (default = 0).}

\item{padj_threshold}{Threshold value for adjusted p-value significance (Defaults to 0.05).}

\item{which_list}{one of c("up_genes","down_genes","up_down_genes","everything"). Select a list of genes to perform the enrichment. Respectively, both up and down regulated genes (up_down_genes), only up regulated genes (up_genes), only down regulated genes (down_genes), or everything allow to load all the three kind of lists separately.}

\item{from_DE_analysis}{Default is TRUE, set FALSE if the lists you want to upload are not from a differential expression analysis.where_files}

\item{files_format}{(Default = NULL). when from_DE_analysis = T it is mandatory to provide the extension of the list of genes to upload.}

\item{where_results}{Specify the folder in which you want to save the outputs. Default is "./". Note: if you are working with RNotebook the default working directory, if not specified, is the folder in which the .Rmd file is saved.}

\item{outfolder}{The name assigned to the folder in which outputs are saved. Default is: "results/". NOTE: please add "/" at the end.}

\item{where_files}{(Default = NULL). When from_DE_analysis = T it is mandatory to provide a path to specify where the list of genes are.}
}
\description{
Function employed for reading all the gene lists on which the enrichment will be performed.
}
