% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGO-package.R, R/autoGO.R
\docType{package}
\name{autoGO}
\alias{autoGO}
\alias{_PACKAGE}
\alias{autoGO-package}
\title{autoGO: Auto-GO: Reproducible, Robust and High Quality Ontology Enrichment Visualizations}
\usage{
autoGO(
  list_of_genes,
  dbs = c("GO_Molecular_Function_2021", "GO_Biological_Process_2021", "KEGG_2021_Human"),
  my_comparison,
  ensembl = FALSE,
  excel = FALSE,
  where_results = "./",
  outfolder = "results/"
)
}
\arguments{
\item{list_of_genes}{it can be a list of dataframe (i.e. the output of read_gene_list()), a vector or a dataframe containing a list of genes or a path to the .txt file.}

\item{dbs}{Databases over which the enrichment will be performed, based on the enrichR libraries. Default are GO_Molecular_Function_2021, GO_Cellular_Component_2021, GO_Biological_Process_2021, KEGG_2021_Human.}

\item{my_comparison}{Name of the comparison the user would like to inspect.}

\item{ensembl}{(Default = FALSE). Set to TRUE if the provided gene list contains Ensembl IDs. A conversion to HGNC will be performed.}

\item{excel}{(Default = FALSE). Set to TRUE if you want to save output tables in .xlsx format.}

\item{where_results}{Specify the folder in which you want to save the outputs. Default is "./". Note: if you are working with RNotebook the default working directory, if not specified, is the folder in which the .Rmd file is saved.}

\item{outfolder}{The name to assign to the folder in which outputs are saved. Default is: "results/". NOTE: please add "/" at the end.}
}
\description{
Auto-GO is a framework that enables automated, high quality Gene Ontology enrichment analyses visualizations. It also features a handy wrapper for Differential Expression analysis around the DESeq2 package (Love et. al 2014) and a handy wrapper around the single-sample implementation of the GSEA algorithm (ssGSEA), useful for low-numerosity transcriptional profiles. The whole framework is structured in different, independent functions, in order to let the user decide which steps of the analysis perform and which plot to produce.

Perform enrichment analysis on all the desired gene lists. This function take advantage of the 'enrichR' package.
}
\author{
\strong{Maintainer}: Isabella Grassucci \email{isabella.grassucci@gmail.com} (\href{https://orcid.org/0000-0002-2484-8570}{ORCID})

Authors:
\itemize{
  \item Eleonora Sperandio \email{eleonora.sperandio@ifo.it} (\href{https://orcid.org/0000-0002-8572-6788}{ORCID})
  \item Lorenzo D'Ambrosio \email{lorenzo.dambrosio@ifo.it} (\href{https://orcid.org/0000-0002-9656-7436}{ORCID})
  \item Matteo Pallocca \email{matteo.pallocca@ifo.it} (\href{https://orcid.org/0000-0002-7756-3579}{ORCID})
  \item Fabio Ticconi \email{fabio.ticconi@gmail.com} (\href{https://orcid.org/0000-0001-8905-6657}{ORCID})
  \item Alice Massacci \email{alice.massacci@ifo.it} (\href{https://orcid.org/0000-0001-7780-809X}{ORCID})
}

}
\keyword{internal}
