% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enrichr_database}
\alias{enrichr_database}
\title{List of databases supported by EnrichR.}
\format{
A data frame with 179 rows and 1 variable:
\describe{
\item{DATASETS}{Identifier of the database to be used for enrichment}
}
}
\usage{
enrichr_database
}
\description{
TODO: write a longer description
}
\keyword{datasets}
